Refactor parseRemark logic and relocate CommandType for cohesion
