+++ src/main/java/seedu/finclient/logic/commands/RemarkUtil/CommandType.java
@@ -1,4 +1,4 @@
-package seedu.finclient.logic.commands;
+package seedu.finclient.logic.commands.RemarkUtil;
 
 /**
  * Represents the context in which a command is being parsed or executed.
