+++ src/main/java/seedu/finclient/logic/parser/EditCommandParser.java
@@ -19,7 +19,7 @@ import java.util.Optional;
 import java.util.Set;
 
 import seedu.finclient.commons.core.index.Index;
-import seedu.finclient.logic.commands.CommandType;
+import seedu.finclient.logic.commands.RemarkUtil.CommandType;
 import seedu.finclient.logic.commands.EditCommand;
 import seedu.finclient.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.finclient.logic.parser.exceptions.ParseException;
